# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "WiSAR Terrain-Aware Range Rings API",
    "version": "1.0.0-draft",
    "summary": "Headless API for Terrain-Aware Range Rings (TARR) and travel-time isochrones.",
...
    "license": {
      "name": "GPL-3.0",
      "identifier": "GPL-3.0-only"
    }
  },
  "servers": [
    {
      "url": "{scheme}://{host}/api/v1",
      "variables": {
        "scheme": {
          "default": "https",
          "enum": [
            "https",
            "http"
          ]
        },
        "host": {
          "default": "wisar.example.org",
          "description": "Hostname (and port) the WiSAR container is published at."
        }
      }
    }
  ],
  "security": [
    {
      "cloudtakBearer": []
    }
  ],
...
  "paths": {
...
    "/profiles": {
      "get": {
        "tags": [
          "Profiles"
        ],
        "operationId": "listProfiles",
        "summary": "List profile datasets",
        "description": "Every installed profile dataset with its categories, eco-region/terrain variants, p25/p50/p75 distances and calibration multipliers. Use this to populate subject pickers.",
        "responses": {
          "200": {
            "description": "Datasets.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileDatasetList"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/NotRegistered"
          },
          "503": {
            "$ref": "#/components/responses/AuthUnavailable"
          }
        }
      }
    },
    "/content/{content_id}": {
      "get": {
        "tags": [
          "Content"
        ],
        "operationId": "getContent",
        "summary": "Get a reference content section",
        "description": "One section of the WiSAR web tool's reference content, read from its static/index.html at request time, so it follows upstream updates. `html` is the section body as an HTML fragment: the web tool's markup and inline styles, with its Close/Got it button, scripts, form controls and inline event handlers removed. The inline styles use CSS custom properties (for example `var(--text-secondary)`); `css_variables` gives the web tool's values for the ones used, so a client can set them on the element it renders into. The colours were chosen for a dark background.\n\nSections: `metadata`, `changelog`, `validation`, `tarr-explainer` (\"What's a TARR?\"), `travel-time-explainer` (\"What's Travel Time?\"), and `scope-note` (the note under the mode choice on the web tool's start screen).\n\nResponses carry an `ETag`; send it back in `If-None-Match` to get `304 Not Modified` when the page has not changed.",
        "parameters": [
          {
            "name": "content_id",
            "in": "path",
            "required": true,
            "description": "Section to return.",
            "schema": {
              "type": "string",
              "enum": [
                "metadata",
                "changelog",
                "validation",
                "tarr-explainer",
                "travel-time-explainer",
                "scope-note"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The section.",
            "headers": {
              "ETag": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Content"
                }
              }
            }
          },
          "304": {
            "description": "Not modified since the ETag sent in If-None-Match."
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/NotRegistered"
          },
          "404": {
            "description": "No section with that id.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "503": {
            "description": "The web tool page is missing, the section was not found in it (renamed upstream), or CloudTAK could not be reached to verify the token.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
    "/tarr/jobs": {
      "post": {
        "tags": [
          "TARR"
        ],
        "operationId": "createTarrJob",
        "summary": "Start a TARR analysis",
        "description": "Queues a TARR analysis. The subject is either a **listed** profile (the server resolves distances from the dataset and applies its calibration) or a **custom** profile with caller-supplied p25/p50/p75. Analysis radius is calibrated p75 + 2 km.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TarrJobRequest"
              },
              "examples": {
                "listed": {
                  "summary": "Listed Koester profile",
                  "value": {
                    "ipp": {
                      "lat": 34.9523,
                      "lon": -111.761
                    },
                    "dataset": "koester",
                    "subject": {
                      "kind": "listed",
                      "category": "Hiker",
                      "eco_region": "Dry",
                      "terrain": "Mountainous"
                    }
                  }
                },
                "custom": {
                  "summary": "Custom (unlisted) profile, miles, uncalibrated",
                  "value": {
                    "ipp": {
                      "lat": 34.9523,
                      "lon": -111.761
                    },
                    "subject": {
                      "kind": "custom",
                      "name": "AZ Hiker (local)",
                      "distances": {
                        "p25": 1.2,
                        "p50": 2.4,
                        "p75": 4.6,
                        "unit": "mi"
                      }
                    },
                    "calibration": "none"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "202": {
            "$ref": "#/components/responses/JobAccepted"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/NotRegistered"
          },
          "422": {
            "$ref": "#/components/responses/Unprocessable"
          },
          "503": {
            "$ref": "#/components/responses/Unavailable"
          }
        }
      }
    },
    "/travel-time/jobs": {
      "post": {
        "tags": [
          "Travel Time"
        ],
        "operationId": "createTravelTimeJob",
        "summary": "Start a travel-time analysis",
        "description": "Queues a travel-time (isochrone) analysis. Travel time per cell = cost-distance ÷ base speed. Analysis radius is max(`min_radius_m`, base speed × longest interval + 2 km).",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TravelTimeJobRequest"
              },
              "examples": {
                "default": {
                  "value": {
                    "ipp": {
                      "lat": 34.9523,
                      "lon": -111.761
                    },
                    "speed": {
                      "value": 2.0,
                      "unit": "mph"
                    },
                    "intervals_hours": [
                      2,
                      4,
                      6,
                      8,
                      10,
                      12
                    ]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "202": {
            "$ref": "#/components/responses/JobAccepted"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/NotRegistered"
          },
          "422": {
            "$ref": "#/components/responses/Unprocessable"
          },
          "503": {
            "$ref": "#/components/responses/Unavailable"
          }
        }
      }
    },
    "/jobs": {
      "get": {
        "tags": [
          "Jobs"
        ],
        "operationId": "listJobs",
        "summary": "List my jobs",
        "description": "Unexpired jobs created by the calling CloudTAK user, newest first.",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/JobStatus"
            },
            "description": "Filter by status."
          },
          {
            "name": "type",
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/JobType"
            },
            "description": "Filter by job type."
          }
        ],
        "responses": {
          "200": {
            "description": "Jobs.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "jobs"
                  ],
                  "properties": {
                    "jobs": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Job"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/NotRegistered"
          },
          "503": {
            "$ref": "#/components/responses/AuthUnavailable"
          }
        }
      }
    },
    "/jobs/{job_id}": {
      "parameters": [
        {
          "$ref": "#/components/parameters/JobId"
        }
      ],
      "get": {
        "tags": [
          "Jobs"
        ],
        "operationId": "getJob",
        "summary": "Get job status and results",
        "description": "Any authenticated user holding the job id can read it, so a job can be shared within an incident.",
        "responses": {
          "200": {
            "description": "The job.",
            "headers": {
              "Retry-After": {
                "schema": {
                  "type": "integer"
                },
                "description": "Suggested seconds before polling again. Present while queued or running."
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Job"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/NotRegistered"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "410": {
            "$ref": "#/components/responses/Gone"
          },
          "503": {
            "$ref": "#/components/responses/AuthUnavailable"
          }
        }
      },
      "delete": {
        "tags": [
          "Jobs"
        ],
        "operationId": "deleteJob",
        "summary": "Cancel or delete a job",
        "description": "Removes a queued or finished job and its outputs. Only the user who created the job may delete it. A running job cannot be interrupted (409).",
        "responses": {
          "204": {
            "description": "Deleted."
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "description": "Caller did not create this job (same user on the same CloudTAK deployment), or the request's Origin is not a registered CloudTAK deployment.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "410": {
            "$ref": "#/components/responses/Gone"
          },
          "409": {
            "description": "Job is running.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "503": {
            "$ref": "#/components/responses/AuthUnavailable"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "cloudtakBearer": {
        "type": "http",
        "scheme": "bearer",
        "description": "CloudTAK user token, verified against GET /api/login on the CloudTAK deployment the request comes from (its Origin), or on the single configured CloudTAK."
      }
    },
    "parameters": {
      "JobId": {
        "name": "job_id",
        "in": "path",
        "required": true,
        "description": "Job id returned when the job was created.",
        "schema": {
          "type": "string",
          "format": "uuid"
        }
      }
    },
    "responses": {
      "JobAccepted": {
        "description": "Job queued.",
        "headers": {
          "Location": {
            "schema": {
              "type": "string",
              "format": "uri-reference"
            },
            "description": "URL of the job."
          },
          "Retry-After": {
            "schema": {
              "type": "integer"
            },
            "description": "Suggested seconds before first poll."
          }
        },
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Job"
            }
          }
        }
      },
      "BadRequest": {
        "description": "Body is not valid JSON or is not an object.",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            }
          }
        }
      },
      "Unprocessable": {
        "description": "Body failed validation. `errors` lists each field.",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            }
          }
        }
      },
      "Unauthorized": {
        "description": "Missing or invalid CloudTAK token.",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            }
          }
        }
      },
      "NotFound": {
        "description": "No such job (or no such output for this job type).",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            }
          }
        }
      },
      "Gone": {
        "description": "Job expired and its outputs were deleted.",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            }
          }
        }
      },
      "AuthUnavailable": {
        "description": "CloudTAK could not be reached to verify the token.",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            }
          }
        }
      },
      "Unavailable": {
        "description": "Queue is full, or CloudTAK could not be reached to verify the token.",
        "headers": {
          "Retry-After": {
            "schema": {
              "type": "integer"
            }
          }
        },
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            }
          }
        }
      },
      "NotRegistered": {
        "description": "The request's Origin is not a CloudTAK deployment this WiSAR serves (WISAR_CLOUDTAK_INSTANCES).",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            }
          }
        }
      }
    },
    "schemas": {
      "Problem": {
        "type": "object",
        "description": "RFC 9457 problem details.",
        "required": [
          "type",
          "title",
          "status"
        ],
        "properties": {
          "type": {
            "type": "string",
            "format": "uri-reference",
            "default": "about:blank"
          },
          "title": {
            "type": "string"
          },
          "status": {
            "type": "integer"
          },
          "detail": {
            "type": "string"
          },
          "instance": {
            "type": "string",
            "format": "uri-reference"
          },
          "errors": {
            "type": "array",
            "description": "Field-level validation errors (422 only).",
            "items": {
              "type": "object",
              "required": [
                "pointer",
                "detail"
              ],
              "properties": {
                "pointer": {
                  "type": "string",
                  "description": "JSON Pointer into the request body, e.g. /subject/distances/p50."
                },
                "detail": {
                  "type": "string"
                }
              }
            }
          }
        }
      },
...
      "JobType": {
        "type": "string",
        "enum": [
          "tarr",
          "travel_time"
        ]
      },
      "JobStatus": {
        "type": "string",
        "enum": [
          "queued",
          "running",
          "succeeded",
          "failed"
        ]
      },
...
      "TarrJobRequest": {
        "type": "object",
        "required": [
          "ipp",
          "subject"
        ],
        "additionalProperties": false,
        "properties": {
          "ipp": {
            "$ref": "#/components/schemas/LatLon"
          },
          "dataset": {
            "type": "string",
            "default": "koester",
            "description": "Dataset id from GET /profiles. Used for listed subjects and for `global` calibration."
          },
          "subject": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/ListedSubject"
              },
              {
                "$ref": "#/components/schemas/CustomSubject"
              }
            ],
            "discriminator": {
              "propertyName": "kind",
              "mapping": {
                "listed": "#/components/schemas/ListedSubject",
                "custom": "#/components/schemas/CustomSubject"
              }
            }
          },
          "calibration": {
            "$ref": "#/components/schemas/Calibration"
          }
        }
      },
...
      "TravelTimeJobRequest": {
        "type": "object",
        "required": [
          "ipp",
          "speed",
          "intervals_hours"
        ],
        "additionalProperties": false,
        "properties": {
          "ipp": {
            "$ref": "#/components/schemas/LatLon"
          },
          "speed": {
            "$ref": "#/components/schemas/Speed"
          },
          "intervals_hours": {
            "type": "array",
            "minItems": 1,
            "maxItems": 12,
            "uniqueItems": true,
            "items": {
              "type": "number",
              "exclusiveMinimum": 0,
              "maximum": 72
            },
            "description": "Isochrone times in hours. Sorted server-side."
          },
          "min_radius_m": {
            "type": "number",
            "minimum": 1000,
            "default": 10000,
            "description": "Minimum analysis radius. The server enlarges it to base speed × longest interval + 2 km."
          }
        }
      },
...
      "Job": {
        "type": "object",
        "required": [
          "id",
          "type",
          "status",
          "created_at",
          "request",
          "links"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "type": {
            "$ref": "#/components/schemas/JobType"
          },
          "status": {
            "$ref": "#/components/schemas/JobStatus"
          },
          "queue_position": {
            "type": [
              "integer",
              "null"
            ],
            "description": "1 = next to run. Null unless queued."
          },
          "owner": {
            "type": "string",
            "description": "CloudTAK username (email) that created the job."
          },
          "instance": {
            "type": [
              "string",
              "null"
            ],
            "description": "Web origin of the CloudTAK deployment the owner signed in to (WISAR_CLOUDTAK_INSTANCES). Null when WiSAR serves a single CloudTAK. Jobs are listed and deleted per owner and deployment."
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "started_at": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "finished_at": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "expires_at": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "request": {
            "type": "object",
            "description": "The request body as received."
          },
          "resolved": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/ResolvedTarr"
              },
              {
                "$ref": "#/components/schemas/ResolvedTravelTime"
              }
            ]
          },
          "result": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/JobResult"
              },
              {
                "type": "null"
              }
            ]
          },
          "outputs": {
            "type": [
              "object",
              "null"
            ],
            "description": "Present when succeeded. Keys are output names.",
            "additionalProperties": {
              "$ref": "#/components/schemas/OutputLink"
            },
            "propertyNames": {
              "enum": [
                "contours.geojson",
                "contours.kml",
                "cost-distance.tif",
                "cost-surface.tif",
                "attractor-score.tif",
                "probability.tif"
              ]
            }
          },
          "error": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/Problem"
              },
              {
                "type": "null"
              }
            ],
            "description": "Present when failed."
          },
          "links": {
            "type": "object",
            "required": [
              "self"
            ],
            "properties": {
              "self": {
                "type": "string",
                "format": "uri-reference"
              }
            }
          }
        }
      },
...
      "ProfileDatasetList": {
        "type": "object",
        "required": [
          "datasets",
          "default_dataset"
        ],
        "properties": {
          "default_dataset": {
            "type": "string"
          },
          "datasets": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProfileDataset"
            }
          }
        }
      },
...
      "Content": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "id",
          "title",
          "html",
          "css_variables",
          "source"
        ],
        "properties": {
          "id": {
            "type": "string",
            "enum": [
              "metadata",
              "changelog",
              "validation",
              "tarr-explainer",
              "travel-time-explainer",
              "scope-note"
            ]
          },
          "title": {
            "type": "string",
            "description": "The section's heading in the web tool.",
            "examples": [
              "Understanding Travel Time analysis"
            ]
          },
          "html": {
            "type": "string",
            "description": "HTML fragment with inline styles. No scripts, form controls or event handlers."
          },
          "css_variables": {
            "type": "object",
            "description": "Values of the CSS custom properties the fragment uses, from the web tool's :root.",
            "additionalProperties": {
              "type": "string"
            },
            "examples": [
              {
                "--text-secondary": "#9498a3",
                "--border": "#2e3138"
              }
            ]
          },
          "source": {
            "type": "object",
            "additionalProperties": false,
            "required": [
              "file",
              "element_id",
              "sha256"
            ],
            "description": "Where the section was read from.",
            "properties": {
              "file": {
                "type": "string",
                "examples": [
                  "static/index.html"
                ]
              },
              "element_id": {
                "type": "string",
                "description": "id of the modal in the web tool page.",
                "examples": [
                  "travelTimeInfoModal"
                ]
              },
              "sha256": {
                "type": "string",
                "pattern": "^[0-9a-f]{64}$",
                "description": "Hash of the page the section came from."
              }
            }
          }
        }
      }
    }
  }
}